[docs] write home page content, remove redundant scripts
